+++ src/ui/colorManager/colorPalette/ColorPalette.h
@@ -9,20 +9,20 @@
 
 class ColorPalette {
   public:
-    ColorPalette(uint32_t contrastColor, uint32_t backgroundColor, std::vector<uint32_t> containerColors);
-    ColorPalette(uint32_t borderColor, uint32_t textColor, uint32_t backgroundColor, std::vector<uint32_t> containerColors);
-    uint32_t getContrastColor();
-    uint32_t getBorderColor();
-    uint32_t getTextColor();
-    uint32_t getBackgroundColor();
-    uint32_t getContainerColor(uint8_t order);
+    ColorPalette(uint16_t contrastColor, uint16_t backgroundColor, std::vector<uint16_t> containerColors);
+    ColorPalette(uint16_t borderColor, uint16_t textColor, uint16_t backgroundColor, std::vector<uint16_t> containerColors);
+    uint16_t getContrastColor();
+    uint16_t getBorderColor();
+    uint16_t getTextColor();
+    uint16_t getBackgroundColor();
+    uint16_t getContainerColor(uint8_t order);
 
   private:
-    uint32_t  _contrastColor,
+    uint16_t  _contrastColor,
               _borderColor,
               _textColor,
               _backgroundColor;
-    std::vector<uint32_t> _containerColors;
+    std::vector<uint16_t> _containerColors;
     bool commonContrastColor;
 
 };
